Data.Bson

Documentation

class ToBson a where

Methods

toBson :: a -> Value

class FromBson a where

Methods

fromBson :: Value -> Parser a

data Parser a

A continuation-based parser type.

parse :: (a -> Parser b) -> a -> Result b

Run a Parser.

parseMaybe :: (a -> Parser b) -> a -> Maybe b

Run a Parser with a Maybe result type.

parseEither :: (a -> Parser b) -> a -> Either String b

Run a Parser with an Either result type.

data RegexOption

Options for ValueRegex, constructors order is important because it's binary representation should be encoded in alphabetical order.

Constructors

RegexOptionCaseInsensitive

RegexOptionLocaleDependent

RegexOptionMultiline

RegexOptionDotall

RegexOptionUnicode

RegexOptionVerbose

type RegexOptions = BitSet RegexOption

data Value

A value is one of the following types of values

Constructors

ValueDouble !Double

ValueString !Text

ValueDocument !Document

ValueArray !Array

ValueBinary !Binary

ValueObjectId !ObjectId

ValueBool !Bool

ValueUtcTime !UTCTime

ValueNull

ValueRegex !Text !RegexOptions

ValueJavascript !Text

ValueJavascriptWithScope !Text !Document

ValueInt32 !Int32

ValueInt64 !Int64

ValueTimestamp !Int64

ValueMin

ValueMax

data Binary

Constructors

BinaryGeneric !ByteString

BinaryFunction !ByteString

BinaryUuid !UUID

BinaryMd5 !ByteString

BinaryUserDefined !ByteString

data ObjectId

Constructors

ObjectId

Fields

objectIdTime :: !Word32

objectIdMachine :: !Word24

objectIdPid :: !Word16

objectIdInc :: !Word24

type Document = HashMap Label Value

type Array = Vector Value

type Label = Text

type Field = (Label, Value)

document :: [Field] -> Document

(!?) :: FromBson a => Document -> Label -> Parser a

Recursively lookup a nested field in a Document.

(=:) :: ToBson a => Label -> a -> Field