Data.Bson

Documentation

class ToBson a where

Methods

toBson :: a -> BsonValue

class FromBson a where

Methods

fromBson :: BsonValue -> Either Text a

data BsonRegexOption

Options for BsonValueRegex, constructors order is important because it's binary representation should be encoded in alphabetical order.

Constructors

BsonRegexOptionCaseInsensitive

BsonRegexOptionLocaleDependent

BsonRegexOptionMultiline

BsonRegexOptionDotall

BsonRegexOptionUnicode

BsonRegexOptionVerbose

type BsonRegexOptions = BitSet BsonRegexOption

data BsonValue

A BSON value is one of the following types of values

Constructors

BsonValueDouble !Double

BsonValueString !Text

BsonValueDocument !BsonDocument

BsonValueArray !BsonArray

BsonValueBinary !BsonBinary

BsonValueObjectId !BsonObjectId

BsonValueBool !Bool

BsonValueUtcTime !UTCTime

BsonValueNull

BsonValueRegex !Text !BsonRegexOptions

BsonValueJavascript !Text

BsonValueJavascriptWithScope !BsonDocument !Text

BsonValueInt32 !Int32

BsonValueInt64 !Int64

BsonValueTimestamp !Int64

BsonValueMin

BsonValueMax

data BsonBinary

Constructors

BsonBinaryGeneric !ByteString

BsonBinaryFunction !ByteString

BsonBinaryUuid !UUID

BsonBinaryMd5 !ByteString

BsonBinaryUserDefined !ByteString

data BsonObjectId

Constructors

BsonObjectId

Fields

bsonObjectIdTime :: !Word32

bsonObjectIdMachine :: !Word24

bsonObjectIdPid :: !Word16

bsonObjectIdInc :: !Word24

type BsonDocument = HashMap BsonLabel BsonValue

type BsonArray = Vector BsonValue

type BsonLabel = Text

type BsonField = (BsonLabel, BsonValue)

document :: [BsonField] -> BsonDocument

(!?) :: FromBson a => BsonDocument -> BsonLabel -> Maybe a

Recursively lookup a nested field in a Document.

(=:) :: ToBson a => BsonLabel -> a -> BsonField